maven: move into java-simple-application
