+++ generators/generate-blueprint/internal/lookup-namespaces.spec.ts
@@ -74,6 +74,7 @@ describe('lookupGeneratorsNamespaces', () => {
   "java-simple-application:code-quality",
   "java-simple-application:graalvm",
   "java-simple-application:jib",
+  "java-simple-application:maven",
   "java-simple-application:openapi-generator",
   "java:bootstrap",
   "java:domain",
@@ -95,7 +96,6 @@ describe('lookupGeneratorsNamespaces', () => {
   "languages",
   "languages:bootstrap",
   "liquibase",
-  "maven",
   "project-name",
   "project-name:bootstrap",
   "react",
