+++ generators/java-simple-application/generators/build-tool/generator.spec.ts
@@ -49,11 +49,11 @@ describe(`generator - ${generator}`, () => {
             buildTool: 'maven',
           })
           .withSkipWritingPriorities()
-          .withMockedGenerators(['jhipster:maven', 'jhipster:gradle']);
+          .withMockedGenerators(['jhipster:java-simple-application:maven', 'jhipster:gradle']);
       });
 
       it('should compose with maven generator', () => {
-        runResult.assertGeneratorComposedOnce('jhipster:maven');
+        runResult.assertGeneratorComposedOnce('jhipster:java-simple-application:maven');
       });
       it('should not compose with others buildTool generators', () => {
         runResult.assertGeneratorNotComposed('jhipster:gradle');
@@ -67,14 +67,14 @@ describe(`generator - ${generator}`, () => {
             buildTool: 'gradle',
           })
           .withSkipWritingPriorities()
-          .withMockedGenerators(['jhipster:maven', 'jhipster:gradle']);
+          .withMockedGenerators(['jhipster:java-simple-application:maven', 'jhipster:gradle']);
       });
 
       it('should compose with gradle generator', () => {
         runResult.assertGeneratorComposedOnce('jhipster:gradle');
       });
       it('should not compose with others buildTool generators', () => {
-        runResult.assertGeneratorNotComposed('jhipster:maven');
+        runResult.assertGeneratorNotComposed('jhipster:java-simple-application:maven');
       });
     });
   });
