+++ generators/java-simple-application/generators/maven/command.ts
@@ -16,15 +16,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { asWritingTask } from '../base-application/support/task-type-inference.ts';
-import type { Application as JavaApplication, Entity as JavaEntity } from '../java/types.d.ts';
+import type { JHipsterCommandDefinition } from '../../../../lib/command/index.ts';
 
-/**
- * Removes server files that where generated in previous JHipster versions and therefore
- * need to be removed.
- */
-export default asWritingTask<JavaEntity, JavaApplication>(function cleanupOldServerFilesTask({ control }) {
-  if (control.isJhipsterVersionLessThan('7.7.1')) {
-    this.removeFile('.mvn/wrapper/MavenWrapperDownloader.java');
-  }
-});
+const command = {
+  configs: {
+    sortMavenPom: {
+      description: 'Sort the pom.xml file',
+      cli: {
+        type: Boolean,
+        hide: true,
+        default: true,
+      },
+      scope: 'generator',
+    },
+  },
+  import: [],
+} as const satisfies JHipsterCommandDefinition;
+
+export default command;
