+++ generators/java-simple-application/generators/maven/constants.ts
