+++ generators/java-simple-application/generators/maven/files.ts
