+++ generators/java-simple-application/generators/maven/index.ts
