+++ generators/java-simple-application/generators/maven/internal/pom-sort.ts
