+++ generators/java-simple-application/generators/maven/internal/xml-store.ts
