+++ generators/java-simple-application/generators/maven/support/dependabot-maven.spec.ts
