+++ generators/java-simple-application/generators/maven/support/index.ts
