+++ generators/java-simple-application/generators/maven/support/pom-file-sort.spec.ts
