+++ generators/java-simple-application/generators/maven/templates/.mvn/jvm.config
