+++ generators/java-simple-application/generators/maven/templates/.prettierignore.jhi.maven.ejs
@@ -17,7 +17,7 @@
  limitations under the License.
 -%>
 <&_ if (!fragment.section) { -&>
-# Generated by jhipster:maven
+# Generated by jhipster:java-simple-application:maven
 target
 .mvn
 <&_ } -&>
