+++ generators/java-simple-application/generators/maven/templates/mvnw
