+++ generators/spring-cloud/generators/gateway/generator.spec.ts
@@ -51,7 +51,7 @@ describe(`generator - ${generator}`, () => {
     it('should compose with generators', () => {
       expect(result.getComposedGenerators()).toMatchInlineSnapshot(`
 [
-  "jhipster:maven",
+  "jhipster:java-simple-application:maven",
 ]
 `);
     });
@@ -78,7 +78,7 @@ describe(`generator - ${generator}`, () => {
     it('should compose with generators', () => {
       expect(result.getComposedGenerators()).toMatchInlineSnapshot(`
 [
-  "jhipster:maven",
+  "jhipster:java-simple-application:maven",
 ]
 `);
     });
@@ -119,7 +119,7 @@ describe(`generator - ${generator}`, () => {
     it('should compose with generators', () => {
       expect(result.getComposedGenerators()).toMatchInlineSnapshot(`
 [
-  "jhipster:maven",
+  "jhipster:java-simple-application:maven",
 ]
 `);
     });
